+++ src/main/java/com/doback/E_rank/interfaces/UsuariosRepository.java
@@ -5,15 +5,15 @@ import java.util.List;
 import java.util.Optional;
 
 public interface UsuariosRepository {
-    public UsuariosModel searchByCode(Long id);
+    public UsuariosModel searchByCode(int id);
 
     public List<UsuariosModel> buscar();
 
     public void addUsuarios(UsuariosModel usuariosModel);
 
-    public void removeUsuarios(Long id);
+    public void removeUsuarios(int id);
 
-    public void updateUsuarios(Long id, UsuariosModel usuariosModel);
+    public void updateUsuarios(int id, UsuariosModel usuariosModel);
 
     public boolean estaVazio();
